refactor: Github API status를 나타내는 element 부분을 별개의 compomnent로 분리
